testmessage: move deserialization related codes into their own module
